requirements.txt
requests==2.25.1
discord.py==1.6.0
PyNaCl==1.4.0
youtube-dl==2021.1.16